+++ deployment/data_handling.py
@@ -67,7 +67,7 @@ def data_handling(graphs_folder):
 		edges = [edge_in, edge_out]
 		edges = np.asarray(edges)
 
-		x = torch.tensor(node_features, dtype=torch.float)
+		x = torch.tensor(node_features, dtype=torch.double)
 		# y = torch.tensor(labels, dtype=torch.float)
 		edge_index = torch.tensor(edges, dtype=torch.long)
 		#     print("X.size: {}, y.size: {}, edge_index.size: {}".format(x.size(), y.size(), edge_index.size()))
